third_party/blink/web_tests/fast/media/matchmedium-query-api-expected.txt

Test the media interface: http://dev.w3.org/csswg/cssom-view/#the-media-interface.

"screen" evaluates to true: PASS
"projection" evaluates to false: PASS
"all and (color)" evaluates to true: PASS
"not projection and (color)" evaluates to true: PASS
"(color)" evaluates to true: PASS
"(color" evaluates to true: PASS
"color" evaluates to false: PASS
"garbage" evaluates to false: PASS
"(min-device-width: 1px)" evaluates to true: PASS
"(min-device-width: 50000px)" evaluates to false: PASS
